requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
discord.py==1.3.3
idna==2.9
multidict==4.7.5
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
urllib3==1.25.8
websockets==8.1
yarl==1.4.2
